% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pol_net_measure.R
\name{pol_net_measure}
\alias{pol_net_measure}
\title{Network-based polarization measure}
\usage{
pol_net_measure(A, party, method)
}
\arguments{
\item{A}{\eqn{N\times N} symmetric weighted adjacency matrix of vote profile correlation.}

\item{party}{\eqn{N\times 1} vector of voter party affiliation.}

\item{method}{Polarization measure.
\itemize{
\item \code{QNewman} Modularity value calculated following Newman (2006).
\item \code{NCut} \eqn{-1\times} Normalized cut calculated following Shi and Malik (2000).
}}
}
\value{
Polarization measure.
}
\description{
Polarization measure calculation from an \eqn{N\times N} vote profile correlation matrix.
}
\examples{
## Generate a vote matrix with 1000 votes in a two party legislature
party <- rbind(matrix(1,50,1),matrix(2,50,1))
vote_data <- pol_simul(party = party, M = 1000, partyMean = c(-1,1), partySD = c(.5,.5))
V <- vote_data$votes

## Sum up vote profile correlation matrices for yes and no votes respectively
A <- net_newman(V)+net_newman(!V)

## Calculate Modularity and - Normalized cut for matrix A
pol_value <- matrix(0,2,1)
for (i in 1:2){
  pol_value[i,1] <- pol_net_measure(A, party, i)
}

}
\references{
Newman, Mark EJ. "Modularity and community structure in networks." \emph{Proceedings of the national academy of sciences} 103.23 (2006): 8577-8582.

Shi, Jianbo, and Jitendra Malik. "Normalized cuts and image segmentation." \emph{IEEE Transactions on pattern analysis and machine intelligence} 22.8 (2000): 888-905.
}
