% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pol_simul.R
\name{pol_simul}
\alias{pol_simul}
\title{Simulate data using the quadratic-normal voting model}
\usage{
pol_simul(party, M, partyMean, partySD)
}
\arguments{
\item{party}{\eqn{N\times 1} vector of voter party affiliation.}

\item{M}{Number of voting items.}

\item{partyMean}{\eqn{P\times 1} vector of party mean ideology for \eqn{P} parties.}

\item{partySD}{\eqn{P\times 1} vector for standard deviation of intra-party ideology.}
}
\value{
\itemize{
\item \code{votes} \eqn{N\times M} binary vote matrix with yes: 1 and no: 0 generated from the quadratic normal voting model (Poole, 2005) with Gaussian noise drawn from \eqn{\mathcal{N}(0,0.2^2)}.
\item \code{ideology} Ideology of legislators used in the simulation.
\item \code{a} \eqn{M\times 1} vector of ideological locations of bill proposals drawn from \eqn{\mathcal{N}(0,1)}.
\item \code{q} \eqn{M\times 1} vector of ideological locations of corresponding status quos drawn from \eqn{\mathcal{N}(0,1)}.
}
}
\description{
Generate simulated data using the quadratic-normal voting model (Poole, 2005) with ideological homogeneity among same party members and ideological distance between different party members.
}
\examples{
## Generate a vote matrix with 1000 votes in a two party legislature consisting of 100 legislators
party <- rbind(matrix(1,50,1),matrix(2,50,1))
vote_data <- pol_simul(party = party, M = 1000, partyMean = c(-1,1), partySD = c(.5,.5))

}
\references{
Poole, Keith T. \emph{Spatial models of parliamentary voting}. Cambridge University Press, 2005.
}
